chore: move alerts to separate VueX module
